Move AttachmentsTest into HintsTest; add param names
